requirements.txt
annotated_types==0.7.0
baml_py==0.82.0
cffi==1.17.1
click==8.1.8
datasets==3.4.1
docstring_parser==0.16
GitPython==3.1.44
hnswlib==0.8.0
huggingface_hub==0.30.1
instructor==1.7.9
kp==0.9.0
langchain_core==0.3.51
langchain_text_splitters==0.3.8
marker==2.1.3
matplotlib==3.10.1
numpy==2.2.4
ollama==0.4.7
openai==1.71.0
pandas==2.2.3
peft==0.15.0
Pillow==11.1.0
pydantic==2.11.2
pyshader==0.7.0
pytest==8.3.5
python-dotenv==1.1.0
PyYAML==6.0.2
Requests==2.32.3
safetensors==0.5.3
seaborn==0.13.2
sentencepiece==0.2.0
setuptools==66.1.1
sphinx_material==0.0.36
tabulate==0.9.0
torch==2.6.0+cu126
tqdm==4.67.1
transformers==4.51.0
trl==0.15.2
typer==0.15.2
typing_extensions==4.13.1
wget==3.2
